Update format to api.js and main.js
Added a logout button and function. So when user is logged in it displays a log out icon instead. When that is clicked the user will be logged out and can be directed to log in again.  Display star rating has been removed and updated into the featuredProducts in api.js